pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[tool.python]
version = ">=3.9,<3.12"

[project]
name = "fraudguard-pro"
version = "1.0.0"
description = "Professional fraud detection dashboard"
requires-python = ">=3.9,<3.12"
dependencies = [
    "Flask>=2.3.0",
    "pandas>=2.0.0,<2.2.0",
    "scikit-learn>=1.3.0,<1.4.0",
    "numpy>=1.23,<1.26",
    "google-generativeai>=0.3.0"
]
